Remove TeplIoErrorInfoBar private class
This was a private class, with the intention to use it later. The code
came from gedit. But in the meantime the gedit code evolved, and in the
end it was a better idea to first refactor the gedit code (to test it in
a real situation).

So look at gedit-io-error-info-bar. It would be a good idea to bring it
again to Tepl, but with the newer implementation (and in that case,
directly use it in gedit).

Note that Tepl still has tepl-io-error-info-bars (with the s suffix)
that is part of the API.