ContainerHelper -> StackHelper
TeplStackHelper will be specific to GtkStack, not GtkContainer.
To help (even more) the implementation of a GtkStackSwitcher.

Other possible names:
TeplStackSignaller
TeplStackNotifier

But TeplStack*Helper* will not only contain a convenience signal, but it
is planned to also add one or more utility functions (for example, as a
todo already: get sorted children, by taking into account their
position).