encoding: re-integrate TeplEncoding into the build system
The TeplEncoding class was public before, but it is now added as a
private class, because some refactorings will be done to change the API
and create a TeplEncoding interface with a TeplEncodingIconv
implementation.

I prefer to take again the old TeplEncoding as a basis, that way I do
refactorings instead of re-creating something from scratch. With
refactorings there is less chance to forget something.