Settings: add SimpleThemeVariant
For apps that don't fully support the automatic switch between light and
dark GTK theme variants (automatic, i.e., external to the app, coming
from the system and changing during the app running).

Currently the preferences dialog in gedit would require some work to
have two settings keys for chosing a light and a dark style scheme from
libgedit-gtksourceview (the settings keys, plus the GUI to configure
that of course).