PanelItem: drop the interface, keep only the GObject class
"Merge" TeplPanelItemSimple (private class) and TeplPanelItem (public
interface) into a single GObject public class called TeplPanelItem.

And add tepl_panel_item_new(), so that adding a TeplPanelItem to a
TeplPanel *will* be done like this (in a further commit):
```
item = tepl_panel_item_new (...);
tepl_panel_add (panel, item);
g_object_unref (item); // or store the item for later use.
```

Because currently tepl_panel_add() is ugly.

-----

This is also to simplify the code. The interface was not really useful
because there is only a single implementation. In this case it doesn't
really make sense to have several implementations.